+++ src/lifish/entities/HauntedStatue.cpp
@@ -15,7 +15,7 @@ HauntedStatue::HauntedStatue(const sf::Vector2f& pos) : Game::Entity(pos) {
 				*addComponent(new Game::Sprite(*this, Game::getAsset("graphics", "haunted_statue.png"),
 				sf::IntRect(0, 0, Game::TILE_SIZE, 2 * Game::TILE_SIZE)))));
 	addComponent(new Game::Collider(*this, Game::Layers::UNBREAKABLES, 
-				sf::Vector2i(Game::TILE_SIZE, Game::TILE_SIZE)));
+				sf::Vector2f(Game::TILE_SIZE, Game::TILE_SIZE)));
 	addComponent(new Game::ZIndexed(*this, Game::Conf::ZIndex::TALL_ENTITIES));
 	addComponent(new Game::Fixed(*this));
